Moved blend ops to their own files
